datatools: add module data
